requirements.txt
streamlit==1.45.1
google-generativeai==0.8.5
pandas==2.2.3
pillow==11.2.1
python-dotenv==1.0.1
plotly==5.19.0
uuid==1.30 